separate btc_utils
