+++ Cargo.toml
@@ -11,13 +11,13 @@ members = [
     "bridge-sdk/bridge-clients/wormhole-bridge-client",
     "bridge-sdk/bridge-clients/btc-bridge-client",
     "bridge-sdk/crypto-utils",
+    "bridge-sdk/btc-utils"
 ]
 
 [workspace.dependencies]
 sha3 = "0.10.8"
 sha2 = "0.10.8"
 bitcoincore-rpc = "0.19.0"
-merkle-tools = { git = "https://github.com/Near-One/btc-light-client-contract", tag = "v0.2.0" }
 bitcoin = "0.32.5"
 borsh = "1.5.1"
 thiserror = "2.0.3"
@@ -33,6 +33,8 @@ cita_trie = "5.0.1"
 rlp = "0.5.2"
 hasher = "0.1.4"
 lazy_static = "1.4.0"
+merkle-tools = { git = "https://github.com/Near-One/btc-light-client-contract", tag = "v0.2.0" }
+minreq = { version = "2.12.0", features = ["https"] }
 near-jsonrpc-client = "0.16.0"
 near-jsonrpc-primitives = "0.29.1"
 near-primitives = "0.29.1"
