+++ bridge-sdk/bridge-clients/btc-bridge-client/Cargo.toml
@@ -7,14 +7,15 @@ edition = "2021"
 
 [dependencies]
 bitcoincore-rpc.workspace = true
-merkle-tools.workspace = true
 bitcoin.workspace = true
 bridge-connector-common = { path = "../../connectors/bridge-connector-common" }
-serde.workspace = true
-serde_json.workspace = true
+merkle-tools.workspace = true
+minreq.workspace = true
 near-crypto.workspace = true
 near-rpc-client = { path = "../../near-rpc-client" }
 near-primitives.workspace = true
+serde.workspace = true
+serde_json.workspace = true
 tokio.workspace = true
 
 [lib]
