+++ bridge-sdk/connectors/omni-connector/Cargo.toml
@@ -21,6 +21,7 @@ solana-sdk.workspace = true
 eth-proof = { path = "../../eth-proof" }
 near-rpc-client = { path = "../../near-rpc-client" }
 crypto-utils = { path = "../../crypto-utils" }
+btc-utils = { path = "../../btc-utils" }
 bridge-connector-common = { path = "../bridge-connector-common" }
 
 near-bridge-client = { path = "../../bridge-clients/near-bridge-client" }
